Decomposed about into sub components - Aboutgdsc component added
