+++ components/About/Aboutgdsc.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const Aboutgdsc = () => {
+    return (
+        <div className="flex flex-col justify-center items-center p-5">
+            <h1 className="bg-gradient-to-r from-googleGreen via-blue-500 to-googleYellow  inline-block text-transparent bg-clip-text font-bold  md:text-[100px] z-20 text-5xl lg:text-[130px] mb-4">
+                GDSC WoW Kerala
+            </h1>
+            <p className="md:px-56 md:text-xl text-sm lg:text-center text-justify flex justify-center items-center px-4 mb-8">
+                GDSC WoW Kerala is a community of students and developers who are passionate about Google technologies. We are a part of Google Developers Student Clubs, a program for students to learn about Google technologies and build solutions for their local communities.
+            </p>
+        </div>
+    );
+}
+
+export default Aboutgdsc;
