+++ components/About/ImageSlider.tsx
