+++ pages/about.tsx
@@ -2,7 +2,8 @@ import Image from "next/image";
 import { Inter } from "next/font/google";
 import NavBar from "@/components/Navbar";
 import Footer from "@/components/Footer";
-import HorizontalImageSlider from "@/components/ImageSlider";
+import HorizontalImageSlider from "@/components/About/ImageSlider";
+import Aboutgdsc from "@/components/About/Aboutgdsc";
 const inter = Inter({ subsets: ["latin"] });
 
 const AboutPage = () => {
@@ -21,14 +22,7 @@ const AboutPage = () => {
   return (
     <div className="flex flex-col overflow-x-hidden">
       <NavBar />
-      <div className="flex flex-col justify-center items-center p-5">
-        <h1 className="text-white font-bold  text-[50px] lg:text-[150px]">
-          GDSC WoW Kerala
-        </h1>
-        <p className="md:px-56 md:text-xl text-sm lg:text-center text-justify flex justify-center items-center px-4 mb-8">
-          GDSC WoW Kerala is a community of students and developers who are passionate about Google technologies. We are a part of Google Developers Student Clubs, a program for students to learn about Google technologies and build solutions for their local communities.
-        </p>
-      </div>
+      <Aboutgdsc />
       <HorizontalImageSlider images={images} />
       <div className="flex flex-col md:flex-row justify-end items-center md:justify-center">
         <button className="h-20 ease-in-out duration-300 z-20 w-64 gradColor font-medium text-md md:text-xl rounded-sm md:my-8 my-4 mx-4">
